Added Piece movement
Added a function before piece_clicked in the Game loop to determine
if the place clicked is a valid move for the active piece. Then
uses an interface for the Movement object to move the active piece
to the new location, and delete the piece that is being landed upon,
if it exists. Then clears the board of highlighting/outlining/etc