Changed UI constructor + misc
Changed the UI constructor to set pointers to all the neccissary
modules, instead of doing it using 'gets' when the functions were
running that needed them.

Also changed is_valid_move to not automatically move the piece, and the
function that uses it now moves the piece instead if is_valid_move returns
true.